+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   * Fix emoticons parsing when using commands
   * Seperate users Mongo collections into users and usergroups
     - usernames are now userids in the users collections
+  * Add stafflist command
 
 1.9.3 / 2015-04-09
 ==================